+++ src/Adapter/Logger.php
@@ -71,7 +71,7 @@ final class Logger implements Adapter
 
     public function all(): Set
     {
-        return $this->root()->files();
+        return Set::of(...$this->root()->files()->toList());
     }
 
     public function root(): Directory